Refactored template directives
Moved the card container specific templates to a templates folder. Moved the generic card templates to a cards folder that is shared with simple cards. Appended template to the class and file names for clarity.
