--- source/components/cardContainer/card/content/index.ts
@@ -1,11 +1,0 @@
-import { CardHeader } from './cardHeader';
-import { ColumnContent } from './columnContent';
-import { CardContent } from './cardContent';
-import { CardFooter } from './cardFooter';
-
-export const CARD_CONTENT_DIRECTIVES: any[] = [CardHeader, ColumnContent, CardContent, CardFooter];
-
-export * from './cardHeader';
-export * from './columnContent';
-export * from './cardContent';
-export * from './cardFooter';
