--- source/components/cardContainer/card/index.ts
@@ -1,5 +1,0 @@
-import { CARD_CONTENT_DIRECTIVES } from './content/index';
-
-export const CARD_DIRECTIVES: any[] = [CARD_CONTENT_DIRECTIVES];
-
-export * from './content/index';
