+++ source/components/cardContainer/index.ts
@@ -1,13 +1,13 @@
-import * as card from './card/index';
 import * as container from './container/index';
 import * as dataSources from './dataSources/index';
 import * as filters from './filters/index';
 import * as paging from './paging/index';
 import * as sorts from './sorts/index';
+import * as templates from './templates/index';
 
-export const CARD_CONTAINER_DIRECTIVES: any[] = [card.CARD_DIRECTIVES, container.CONTAINER_DIRECTIVES, filters.FILTER_DIRECTIVES];
+export const CARD_CONTAINER_DIRECTIVES: any[] = [container.CONTAINER_DIRECTIVES, filters.FILTER_DIRECTIVES, templates.TEMPLATE_DIRECTIVES];
 
-export { card, container, dataSources, filters, paging, sorts };
+export { container, dataSources, filters, paging, sorts, templates };
 
 // card
 export * from './cardContainer';
