+++ source/components/cardContainer/templates/columnContent.template.ts
@@ -1,7 +1,7 @@
 import { Directive, TemplateRef, Input } from '@angular/core';
 
 @Directive({ selector: '[rlColumnContent]' })
-export class ColumnContent {
+export class ColumnContentTemplate {
 	@Input() rlColumnContentName: string;
 
 	get name(): string {
