+++ source/components/cards/cardContent.template.ts
@@ -1,6 +1,6 @@
 import { Directive, TemplateRef } from '@angular/core';
 
-@Directive({ selector: '[rlCardFooter]' })
-export class CardFooter {
+@Directive({ selector: '[rlCardContent]' })
+export class CardContentTemplate {
 	constructor(public template: TemplateRef<any>) {	}
 }
