+++ source/components/cards/cardFooter.template.ts
@@ -1,6 +1,6 @@
 import { Directive, TemplateRef } from '@angular/core';
 
-@Directive({ selector: '[rlCardContent]' })
-export class CardContent {
+@Directive({ selector: '[rlCardFooter]' })
+export class CardFooterTemplate {
 	constructor(public template: TemplateRef<any>) {	}
 }
