+++ source/components/cards/cardHeader.template.ts
@@ -1,6 +1,6 @@
 import { Directive, TemplateRef } from '@angular/core';
 
 @Directive({ selector: '[rlCardHeader]' })
-export class CardHeader {
+export class CardHeaderTemplate {
 	constructor(public template: TemplateRef<any>) {	}
 }
