+++ source/components/cards/index.ts
@@ -1,0 +1,9 @@
+import { CardHeaderTemplate } from './cardHeader.template';
+import { CardContentTemplate } from './cardContent.template';
+import { CardFooterTemplate } from './cardFooter.template';
+
+export const CARD_TEMPLATE_DIRECTIVES: any[] = [CardHeaderTemplate, CardContentTemplate, CardFooterTemplate];
+
+export * from './cardHeader.template';
+export * from './cardContent.template';
+export * from './cardFooter.template';
